pip_requirements.txt
absl-py==1.3.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cycler==0.11.0
Cython==0.29.32
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
decorator==4.4.2
dill==0.3.4
dlib==19.24.0
dominate==2.8.0
focal-loss-torch==0.1.2
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.14.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.3
grpcio==1.48.2
httplib2==0.22.0
idna==3.4
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mypath==0.1
networkx==2.5.1
ninja==1.11.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.5.64
pandas==1.1.5
patsy==0.5.2
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1630696616009/work
pmdarima==1.8.4
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
statsmodels==0.12.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchaudio==0.10.1
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.11.2
tqdm==4.64.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.0.3
wget==3.2
zipp==3.6.0
